+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- Added `@computedTree` decorator for reactively deriving an alternative view of an original tree that is attached as a property to the original tree to support tree traversal functions, contexts, references, etc. (see "Computed Trees" section in the docs).
+
 ## 0.65.0
 
 - Fixed a `Maximum call stack size exceeded` getting triggered when `onAttachedToRootStore` was defined and many nodes were added at once.